Store IDs in dc_array as u32
This changes dc_array_get_raw API

+++ deltachat-ffi/deltachat.h
@@ -387,7 +387,7 @@ char*            dc_array_get_marker         (const dc_array_t*, size_t index);
 int              dc_array_is_independent     (const dc_array_t*, size_t index);
 
 int              dc_array_search_id          (const dc_array_t*, uint32_t needle, size_t* indx);
-const uintptr_t* dc_array_get_raw            (const dc_array_t*);
+const uint32_t*  dc_array_get_raw            (const dc_array_t*);
 
 
 /**
